Break Dashboard component into small carousel components
